+++ pybop/costs/_likelihoods.py
@@ -39,20 +39,19 @@ class GaussianLogLikelihoodKnownSigma(BaseLikelihood):
         self._offset = -0.5 * self.n_time_data * np.log(2 * np.pi * self.sigma2)
         self._multip = -1 / (2.0 * self.sigma2)
 
-    def _evaluate(self, inputs: Inputs, grad: Union[None, np.ndarray] = None) -> float:
+    def _evaluate(self, inputs: Inputs) -> float:
         """
         Evaluates the Gaussian log-likelihood for the given parameters with known sigma.
         """
-        y = self.problem.evaluate(inputs)
-
-        if not self.verify_prediction(y):
+        if not self.verify_prediction(self.y):
             return -np.inf
 
         e = np.asarray(
             [
                 np.sum(
                     self._offset
-                    + self._multip * np.sum((self._target[signal] - y[signal]) ** 2.0)
+                    + self._multip
+                    * np.sum((self._target[signal] - self.y[signal]) ** 2.0)
                 )
                 for signal in self.signal
             ]
@@ -62,17 +61,17 @@ class GaussianLogLikelihoodKnownSigma(BaseLikelihood):
 
     def _evaluateS1(self, inputs: Inputs) -> tuple[float, np.ndarray]:
         """
-        Calls the problem.evaluateS1 method and calculates the log-likelihood and gradient.
+        Calculates the log-likelihood and gradient.
         """
-        y, dy = self.problem.evaluateS1(inputs)
-
-        if not self.verify_prediction(y):
+        if not self.verify_prediction(self.y):
             return -np.inf, -self._de * np.ones(self.n_parameters)
 
         likelihood = self._evaluate(inputs)
 
-        r = np.asarray([self._target[signal] - y[signal] for signal in self.signal])
-        dl = np.sum((np.sum((r * dy.T), axis=2) / self.sigma2), axis=1)
+        r = np.asarray(
+            [self._target[signal] - self.y[signal] for signal in self.signal]
+        )
+        dl = np.sum((np.sum((r * self.dy.T), axis=2) / self.sigma2), axis=1)
 
         return likelihood, dl
 
@@ -117,6 +116,7 @@ class GaussianLogLikelihood(BaseLikelihood):
         super().__init__(problem)
         self._dsigma_scale = dsigma_scale
         self._logpi = -0.5 * self.n_time_data * np.log(2 * np.pi)
+        self._has_separable_problem = False
 
         self.sigma = Parameters()
         self._add_sigma_parameters(sigma0)
@@ -168,7 +168,7 @@ class GaussianLogLikelihood(BaseLikelihood):
             raise ValueError("dsigma_scale must be non-negative")
         self._dsigma_scale = new_value
 
-    def _evaluate(self, inputs: Inputs, grad: Union[None, np.ndarray] = None) -> float:
+    def _evaluate(self, inputs: Inputs) -> float:
         """
         Evaluates the Gaussian log-likelihood for the given parameters.
 
@@ -189,8 +189,8 @@ class GaussianLogLikelihood(BaseLikelihood):
         if np.any(sigma <= 0):
             return -np.inf
 
-        y = self.problem.evaluate(self.problem.parameters.as_dict())
-        if not self.verify_prediction(y):
+        self.y = self.problem.evaluate(self.problem.parameters.as_dict())
+        if not self.verify_prediction(self.y):
             return -np.inf
 
         e = np.asarray(
@@ -198,7 +198,7 @@ class GaussianLogLikelihood(BaseLikelihood):
                 np.sum(
                     self._logpi
                     - self.n_time_data * np.log(sigma)
-                    - np.sum((self._target[signal] - y[signal]) ** 2.0)
+                    - np.sum((self._target[signal] - self.y[signal]) ** 2.0)
                     / (2.0 * sigma**2.0)
                 )
                 for signal in self.signal
@@ -209,7 +209,7 @@ class GaussianLogLikelihood(BaseLikelihood):
 
     def _evaluateS1(self, inputs: Inputs) -> tuple[float, np.ndarray]:
         """
-        Calls the problem.evaluateS1 method and calculates the log-likelihood.
+        Calculates the log-likelihood and sensitivities.
 
         Parameters
         ----------
@@ -227,14 +227,16 @@ class GaussianLogLikelihood(BaseLikelihood):
         if np.any(sigma <= 0):
             return -np.inf, -self._de * np.ones(self.n_parameters)
 
-        y, dy = self.problem.evaluateS1(self.problem.parameters.as_dict())
-        if not self.verify_prediction(y):
+        self.y, self.dy = self.problem.evaluateS1(self.problem.parameters.as_dict())
+        if not self.verify_prediction(self.y):
             return -np.inf, -self._de * np.ones(self.n_parameters)
 
         likelihood = self._evaluate(inputs)
 
-        r = np.asarray([self._target[signal] - y[signal] for signal in self.signal])
-        dl = np.sum((np.sum((r * dy.T), axis=2) / (sigma**2.0)), axis=1)
+        r = np.asarray(
+            [self._target[signal] - self.y[signal] for signal in self.signal]
+        )
+        dl = np.sum((np.sum((r * self.dy.T), axis=2) / (sigma**2.0)), axis=1)
         dsigma = (
             -self.n_time_data / sigma + np.sum(r**2.0, axis=1) / (sigma**3.0)
         ) / self._dsigma_scale
@@ -276,7 +278,10 @@ class MAP(BaseLikelihood):
         ):
             raise ValueError(f"{self.likelihood} must be a subclass of BaseLikelihood")
 
-    def _evaluate(self, inputs: Inputs, grad=None) -> float:
+        self.parameters = self.likelihood.parameters
+        self._has_separable_problem = self.likelihood._has_separable_problem
+
+    def _evaluate(self, inputs: Inputs) -> float:
         """
         Calculate the maximum a posteriori cost for a given set of parameters.
 
@@ -284,9 +289,6 @@ class MAP(BaseLikelihood):
         ----------
         inputs : Inputs
             The parameters for which to evaluate the cost.
-        grad : array-like, optional
-            An array to store the gradient of the cost function with respect
-            to the parameters.
 
         Returns
         -------
@@ -300,7 +302,10 @@ class MAP(BaseLikelihood):
         if not np.isfinite(log_prior).any():
             return -np.inf
 
-        log_likelihood = self.likelihood._evaluate(inputs)
+        if self._has_separable_problem:
+            self.likelihood.y = self.y
+        log_likelihood = self.likelihood.evaluate(inputs)
+
         posterior = log_likelihood + log_prior
         return posterior
 
@@ -331,7 +336,9 @@ class MAP(BaseLikelihood):
         if not np.isfinite(log_prior).any():
             return -np.inf, -self._de * np.ones(self.n_parameters)
 
-        log_likelihood, dl = self.likelihood._evaluateS1(inputs)
+        if self._has_separable_problem:
+            self.likelihood.y, self.likelihood.dy = (self.y, self.dy)
+        log_likelihood, dl = self.likelihood.evaluateS1(inputs)
 
         # Compute a finite difference approximation of the gradient of the log prior
         delta = self.parameters.initial_value() * self.gradient_step